Fix type errora and make signup page work
