+++ app/web/pages/signup.tsx
@@ -1,6 +1,6 @@
 import { appGetLayout } from "components/AppRoute";
 import Signup from "features/auth/signup/Signup";
-import { AUTH, GLOBAL } from "i18n/namespaces";
+import { AUTH, GLOBAL, LANDING } from "i18n/namespaces";
 import { GetStaticProps } from "next";
 import nextI18nextConfig from "next-i18next.config";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
@@ -9,7 +9,7 @@ export const getStaticProps: GetStaticProps = async ({ locale }) => ({
   props: {
     ...(await serverSideTranslations(
       locale ?? "en",
-      [AUTH, GLOBAL],
+      [AUTH, GLOBAL, LANDING],
       nextI18nextConfig,
     )),
   },
